Fixed calc_test.py
