+++ openquake/risklib/tests/scientific_test.py
@@ -551,137 +551,6 @@ class LossMapMatrixTest(unittest.TestCase):
             scientific.loss_map_matrix([0.55, 0.5], self.curves))
 
 
-class AssetStatisticsTestCase(unittest.TestCase):
-    BASE_EXPECTED_POES = numpy.linspace(1, 0, 11)
-
-    def setUp(self):
-        self.losses = numpy.linspace(0, 1, 11)
-
-    # fake post_processing module singleton
-    class post_processing(object):
-        @staticmethod
-        def mean_curve(_curve_poes, _weights):
-            return (AssetStatisticsTestCase.BASE_EXPECTED_POES)
-
-        @staticmethod
-        def weighted_quantile_curve(_curve_poes, _weights, quantile):
-            return -AssetStatisticsTestCase.BASE_EXPECTED_POES * quantile
-
-        @staticmethod
-        def quantile_curve(_curve_poes, quantile):
-            return AssetStatisticsTestCase.BASE_EXPECTED_POES * quantile
-
-    def test_compute_stats_no_quantiles_no_poes(self):
-        (mean_curve, mean_maps, quantile_curves, quantile_maps) = (
-            scientific.asset_statistics(
-                self.losses, mock.Mock(), [],
-                [None], [], self.post_processing))
-
-        numpy.testing.assert_allclose(mean_curve,
-                                      (self.losses, self.BASE_EXPECTED_POES))
-
-        self.assertEqual(0, quantile_curves.size)
-        self.assertEqual(0, mean_maps.size)
-        self.assertEqual(0, quantile_maps.size)
-
-    def test_compute_stats_quantiles_weighted(self):
-        (mean_curve, mean_maps, quantile_curves, quantile_maps) = (
-            scientific.asset_statistics(
-                self.losses, mock.Mock(),
-                quantiles=[0.1, 0.2],
-                poes=[],
-                weights=[0.1, 0.2],
-                post_processing=self.post_processing))
-
-        numpy.testing.assert_allclose(
-            mean_curve, (self.losses, self.BASE_EXPECTED_POES))
-
-        q1, q2 = quantile_curves
-        numpy.testing.assert_allclose(
-            q1, (self.losses, -self.BASE_EXPECTED_POES * 0.1))
-        numpy.testing.assert_allclose(
-            q2, (self.losses, -self.BASE_EXPECTED_POES * 0.2))
-
-        self.assertEqual(0, mean_maps.size)
-        self.assertEqual(0, quantile_maps.size)
-
-    def test_compute_stats_quantiles_montecarlo(self):
-        (mean_curve, mean_maps, quantile_curves, quantile_maps) = (
-            scientific.asset_statistics(
-                self.losses, mock.Mock(),
-                quantiles=[0.1, 0.2],
-                poes=[],
-                weights=[None, None],
-                post_processing=self.post_processing))
-
-        numpy.testing.assert_allclose(
-            mean_curve, (self.losses, self.BASE_EXPECTED_POES))
-
-        q1, q2 = quantile_curves
-        numpy.testing.assert_allclose(
-            q1, (self.losses, self.BASE_EXPECTED_POES * 0.1))
-        numpy.testing.assert_allclose(
-            q2, (self.losses, self.BASE_EXPECTED_POES * 0.2))
-
-        self.assertEqual(0, mean_maps.size)
-        self.assertEqual(0, quantile_maps.size)
-
-    def test_compute_stats_quantile_poes(self):
-        (mean_curve, mean_map, quantile_curves, quantile_maps) = (
-            scientific.asset_statistics(
-                self.losses, mock.Mock(),
-                quantiles=[0.1, 0.2],
-                poes=[0.2, 0.8],
-                weights=[None],
-                post_processing=self.post_processing))
-
-        numpy.testing.assert_allclose(
-            mean_curve, (self.losses, self.BASE_EXPECTED_POES))
-        q1, q2 = quantile_curves
-        numpy.testing.assert_allclose(
-            q1, (self.losses, self.BASE_EXPECTED_POES * 0.1))
-        numpy.testing.assert_allclose(
-            q2, (self.losses, self.BASE_EXPECTED_POES * 0.2))
-
-        numpy.testing.assert_allclose(mean_map, [0.8, 0.2])
-
-        numpy.testing.assert_allclose(quantile_maps, numpy.zeros((2, 2)))
-
-    def test_exposure(self):
-        resolution = 10
-
-        # testing exposure_statistics with arrays of different shapes
-        for quantile_nr, poe_nr, asset_nr in itertools.product(
-                range(3), range(3), range(1, 4)):
-            with mock.patch(
-                    'openquake.risklib.scientific.asset_statistics') as m:
-                m.return_value = (numpy.empty((2, resolution)),
-                                  numpy.empty(poe_nr),
-                                  numpy.empty((quantile_nr, 2, resolution)),
-                                  numpy.empty((quantile_nr, poe_nr)))
-
-                loss_curves = numpy.empty((asset_nr, 2, resolution))
-
-                (mean_curves, mean_average_losses, mean_maps,
-                 quantile_curves, quantile_average_losses, quantile_maps) = (
-                    scientific.exposure_statistics(
-                        loss_curves,
-                        numpy.empty(poe_nr),
-                        numpy.empty(asset_nr),
-                        numpy.empty(quantile_nr),
-                        mock.Mock()))
-
-                self.assertEqual((asset_nr, 2, resolution), mean_curves.shape)
-                self.assertEqual((asset_nr, ), mean_average_losses.shape)
-                self.assertEqual((poe_nr, asset_nr), mean_maps.shape)
-                self.assertEqual((quantile_nr, asset_nr, 2, resolution),
-                                 quantile_curves.shape)
-                self.assertEqual((quantile_nr, asset_nr),
-                                 quantile_average_losses.shape)
-                self.assertEqual((quantile_nr, poe_nr, asset_nr),
-                                 quantile_maps.shape)
-
-
 class ClassicalDamageTestCase(unittest.TestCase):
     def test_discrete(self):
         hazard_imls = [0.05, 0.2, 0.4, 0.6, 0.8, 1, 1.2, 1.4]
